+++ components/calendly_v2/actions/get-event/get-event.mjs
@@ -5,27 +5,28 @@ export default {
   key: "calendly_v2-get-event",
   name: "Get Event",
   description: "Gets information about an Event associated with a URI.",
-  version: "0.1.1",
+  version: "0.1.2",
   type: "action",
   props: {
     calendly_v2: {
       type: "app",
       app: "calendly_v2",
     },
-    event_uuid: {
+    event_url: {
       type: "string",
-      description: "The event's unique identifier.",
+      description: "The URL of the event to retrieve information about. If you are using the Calendly webhook, you would use *{{steps.trigger.event.payload.event}}*",
     },
   },
   async run({ $ }) {
-  //See the API docs: https://developer.calendly.com/docs/api-docs/reference/calendly-api/openapi.yaml/paths/~1scheduled_events~1%7Buuid%7D/get
-
-    if (!this.event_uuid) {
-      throw new Error("Must provide event_uuid parameter.");
+    if (!this.event_url) {
+      throw new Error("Must provide event_url parameter.");
     }
 
+    const url = new URL(this.event_url);
+    const event_uuid = url.pathname.split("/").pop();
+
     return await axios($, {
-      url: `https://api.calendly.com/scheduled_events/${this.event_uuid}`,
+      url: `https://api.calendly.com/scheduled_events/${event_uuid}`,
       headers: {
         Authorization: `Bearer ${this.calendly_v2.$auth.oauth_access_token}`,
       },
